requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements/test.txt requirements/test.in
#
asgiref==3.9.1
    # via django
certifi==2025.10.5
    # via requests
cffi==2.0.0
    # via pynacl
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   code-annotations
    #   edx-django-utils
code-annotations==2.3.0
    # via -r requirements/test.in
coverage[toml]==7.10.6
    # via pytest-cov
ddt==1.7.2
    # via -r requirements/test.in
deepdiff==8.6.1
    # via -r requirements/test.in
django==4.2.25
    # via
    #   -c https:/raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.in
    #   django-config-models
    #   django-crum
    #   django-filter
    #   django-mysql
    #   django-simple-history
    #   django-waffle
    #   djangorestframework
    #   edx-django-utils
    #   jsonfield
django-config-models==2.9.0
    # via -r requirements/test.in
django-crum==0.7.9
    # via edx-django-utils
django-filter==25.1
    # via -r requirements/test.in
django-mysql==4.18.0
    # via -r requirements/test.in
django-simple-history==3.10.1
    # via -r requirements/test.in
django-waffle==5.0.0
    # via edx-django-utils
djangorestframework==3.16.1
    # via
    #   -r requirements/test.in
    #   django-config-models
dnspython==2.8.0
    # via pymongo
edx-django-utils==8.0.0
    # via django-config-models
edx-opaque-keys[django]==3.0.0
    # via -r requirements/test.in
idna==3.11
    # via requests
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via code-annotations
jsonfield==3.2.0
    # via -r requirements/test.in
markupsafe==3.0.2
    # via jinja2
numpy==2.3.3
    # via -r requirements/test.in
openedx-atlas==0.7.0
    # via -r requirements/base.in
orderly-set==5.5.0
    # via deepdiff
packaging==25.0
    # via pytest
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
psutil==7.0.0
    # via edx-django-utils
pycparser==2.23
    # via cffi
pygments==2.19.2
    # via pytest
pymongo==4.15.3
    # via edx-opaque-keys
pynacl==1.6.0
    # via edx-django-utils
pytest==8.4.2
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==7.0.0
    # via -r requirements/test.in
pytest-django==4.11.1
    # via -r requirements/test.in
python-dateutil==2.9.0.post0
    # via -r requirements/test.in
python-slugify==8.0.4
    # via code-annotations
pyyaml==6.0.2
    # via code-annotations
requests==2.32.5
    # via -r requirements/test.in
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
stevedore==5.5.0
    # via
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
types-python-dateutil==2.9.0.20250822
    # via -r requirements/test.in
types-requests==2.32.4.20250913
    # via -r requirements/test.in
typing-extensions==4.15.0
    # via edx-opaque-keys
urllib3==2.5.0
    # via
    #   requests
    #   types-requests
